+++ language/en/admin/settings.php
@@ -1,27 +1,26 @@
 <?php
 
-$lang['adm_opt_title']         = "Configuration du serveur";
-$lang['adm_opt_game_settings'] = "Param&egrave;tre du jeu";
-$lang['adm_opt_game_name']     = "Nom du jeu";
-$lang['adm_opt_game_gspeed']   = "Vitesse de jeu<br>(normal = 2500)";
-$lang['adm_opt_game_fspeed']   = "Vitesse des flottes<br>(normal = 2500)";
-$lang['adm_opt_game_pspeed']   = "Vitesse de production<br>(normal = 1)";
-$lang['adm_opt_game_forum']    = "Adresse du forum";
+$lang['adm_opt_title']         = "Server configuration";
+$lang['adm_opt_game_settings'] = "Game settings";
+$lang['adm_opt_game_name']     = "Game name";
+$lang['adm_opt_game_gspeed']   = "Gameplay speed<br>(normal = 2500)";
+$lang['adm_opt_game_fspeed']   = "Fleets speed<br>(normal = 2500)";
+$lang['adm_opt_game_pspeed']   = "Production speed<br>(normal = 1)";
+$lang['adm_opt_urlaubus_mode'] = "Disable 24h vacation mode";
+$lang['adm_opt_game_forum']    = "Forum URL";
 $lang['adm_opt_game_copyrigh'] = "Copyright";
-$lang['adm_opt_game_online']   = "On-line Status";
-$lang['adm_opt_game_offreaso'] = "Raison de la fermeture";
-$lang['adm_opt_plan_settings'] = "Param&egrave;tre planete";
-$lang['adm_opt_plan_initial']  = "Taille de la<br>plan&egrave;te m&egrave;re";
-$lang['adm_opt_plan_base_inc'] = "Revenu de base en ";
-$lang['adm_opt_game_debugmod'] = "Mode debugage";
-$lang['adm_opt_game_oth_info'] = "Autres infos";
-$lang['adm_opt_game_oth_news'] = "Afficher cadre de News";
-$lang['adm_opt_game_oth_chat'] = "Afficher Chat externe";
-$lang['adm_opt_game_oth_adds'] = "ADS GOOGLE";
-$lang['adm_opt_game_oth_lstc'] = "Derni&egrave;re coordonn&eacute;es de plan&egrave;te m&egrave;re";
-$lang['adm_opt_game_oth_gala'] = "Galaxie";
-$lang['adm_opt_game_oth_syst'] = "Systeme";
-$lang['adm_opt_game_oth_plan'] = "Plan&egrave;te";
-$lang['adm_opt_btn_save']      = "Sauvegarder";
-
-?>
+$lang['adm_opt_game_online']   = "Game online?";
+$lang['adm_opt_game_offreaso'] = "Maintenance message";
+$lang['adm_opt_plan_settings'] = "Planet settings";
+$lang['adm_opt_plan_initial']  = "Size of the<br>mother planet";
+$lang['adm_opt_plan_fields']   = "fields";
+$lang['adm_opt_plan_base_inc'] = "Basic income of ";
+$lang['adm_opt_plan_per_hour'] = "per hour";
+$lang['adm_opt_game_debugmod'] = "Debug mode";
+$lang['adm_opt_game_oth_info'] = "Other info";
+$lang['adm_opt_game_oth_news'] = "Show news frame";
+$lang['adm_opt_game_oth_chat'] = "Show external chat";
+$lang['adm_opt_game_oth_adds'] = "Google Ads";
+$lang['adm_opt_btn_save']      = "Save settings";
+$lang['adm_opt_saved_title']   = "Success";
+$lang['adm_opt_save_success']  = "Configuration has been saved successfully";